+++ src/components/LoadingIsBusy.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+function LoadingIsBusy(props) {
+	return (
+		<div class='LoadingIsBusy'>
+			<div class='inner one'></div>
+			<div class='inner two'></div>
+			<div class='inner three'></div>
+		</div>
+	);
+}
+
+export default LoadingIsBusy;
