+++ src/components/LoadingLogo.jsx
@@ -1,0 +1,13 @@
+import React, { useEffect } from "react";
+import LoadingSVG from "../assets/LoadingSVG";
+
+function LoadingLogo(props) {
+	useEffect(() => {}, []);
+	return (
+		<div className='loading-logo'>
+			<LoadingSVG />
+		</div>
+	);
+}
+
+export default LoadingLogo;
